requirements.txt
flask==3.0.0
openai==1.3.8
httpx==0.25.2
playwright==1.40.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
PyPDF2==3.0.1
matplotlib==3.8.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
gunicorn==21.2.0
